[MISC] Move LocalStoreService to data layer
